test: decode 테스트 케이스 추가, refactor: 파일 분리
